initial clean up for website redesign
