--- src/components/IndexPage/builds.js
@@ -1,53 +1,0 @@
-import React from "react"
-import PropTypes from "prop-types"
-import styled from "styled-components"
-
-import Project from "./project.js"
-import { HeaderLineBelow } from "styles/components/Header"
-import Section from "styles/components/Section"
-
-const ProjectContainer = styled.div`
-  display: flex;
-  flex-wrap: wrap;
-  align-items: center;
-  justify-content: center;
-
-  @media (min-width: 1000px) {
-    justify-content: space-between;
-  }
-`
-
-const Builds = ({ title, projects }) => (
-  <Section>
-    <HeaderLineBelow>
-      <span>{title}</span>
-    </HeaderLineBelow>
-    <ProjectContainer>
-      {projects.map(p => (
-        <Project
-          title={p.title}
-          tags={p.tags}
-          backgroundImage={p.image}
-          description={p.description}
-          github={p.gitLink}
-          key={p.title}
-        />
-      ))}
-    </ProjectContainer>
-  </Section>
-)
-
-Builds.propTypes = {
-  title: PropTypes.string.isRequired,
-  projects: PropTypes.arrayOf(
-    PropTypes.exact({
-      title: PropTypes.string.isRequired,
-      tags: PropTypes.array.isRequired,
-      image: PropTypes.object.isRequired,
-      description: PropTypes.string.isRequired,
-      gitLink: PropTypes.string.isRequired,
-    })
-  ).isRequired,
-}
-
-export default Builds
