+++ src/content/hero/hero.json
@@ -1,4 +1,4 @@
 {
   "title": "Unleash the Power of Software for Researchers and Students",
-  "background": "./stairs-down.png"
+  "background": "./landing-page.png"
 }
