+++ src/pages/about.js
@@ -1,0 +1,8 @@
+import React from "react"
+import { graphql } from "gatsby"
+
+const AboutPage = ({ data }) => {
+  return <span>About page works!</span>
+}
+
+export default AboutPage
