+++ src/pages/portfolio.js
@@ -1,0 +1,8 @@
+import React from "react"
+import { graphql } from "gatsby"
+
+const PortfolioPage = ({ data }) => {
+  return <span>Portfolio page works!</span>
+}
+
+export default PortfolioPage
